Extract A-not-A and modal polar question detectors
